Provide OppijaNumeroRekisteri to SuoritusJDBCActor
This requires a kludge in ScalatraBootstrap (and results
in the ScalatraBootstrap reference to be passed all the way to
SuoritusJDBCActor via closure), but has the benefit of not
requiring to pass the reference over the actor call, and
reduces the plumbing code on that way.

Next up: using the code in deduplication query.