SEC-68 : Fix single organisation retrieval
It had the nasty bug of adding the organisation in
cache without its children. Unfortunately, even
restarting Suoritusrekisteri would not fix the issue,
as the organisations already present in cache would
not be loaded again.

It would have been simpler to use the hierarchy returning
API for the single organisation fetch, but unfortunately
the single organisation API seems to be the only one that
accepts both oids and koodis...