BUG-1990 : Use downloaded content directly
There seems to be a race condition if a recently
put file is read immediately after putting it.
We can just read the file contents in memory and
give separate handles to the in-memory array for
storing to S3 for debugging and for parsing and
storing to database.

Even with a batch size of 50 000, the size of each
zip file stayed under six megabytes, so buffering
them in memory one at a time should not pose problems.