BUG-1986 automated test and implementation
YtlActor changed to support ask pattern. For that
we needed to pass the reference to sender to YoSuoritusUpdateActor
so that it is able to respond to ask pattern's temporary actor

YtlIntegrations: persistKokelas modified to return future, and
sync methods are modified to wait for future to finish. Only when
everything is finished, we log the success log message.

Unit tests for YtlActor and YtlIntegration are enhanced to be able to
really test the feature. FailureEmailSender class had to be
introduced to be able to inject mock and verify whether
failure emails are sent when needed.

Configurable timeout for persistKokelas and sync()

YtlResources test corrected

Dummy YtlActor follows the ask pattern.