BUG-1986 Do not use await, change YtlActor into a class
It is enough to have timeout for individual kokelas

Make sure temporary actors are cleaned
Added the callback to kill temporary actors at the end
of operation.

Note that test logs might be confusing because things are
happening in multiple threads.

Added waiting time to test case

It seems that the loop to read from database
might end too early