Proof that archive happens within transaction
If not specified otherwise, all db operations done within
a db function will happen in single transaction. This means
that the archive batch is done in single transaction. By default
the batch size is 1000, which means that if one of 1000 rows is
not successfullly written to the archive table, the whole batch will
fail.

Note that currently the transaction will fail also if the row already
exists in the archive table. Unique key is (resource_id, inserted).
This is done on purpose, the situation has to be manually resolved, whether
to update existing archive or ignore (and remove the original).