Change the retry logic and refactor
Got rid of ReTryActor and implement retry logic in simpler
way within YtlKokelasPersister.

Simplify the code to handle the errors when waiting for multiple
Futures. There is no need to return boolean from future, if we
are not intersted in the result. Also there is no need to manually
check whether all Futures have completed succesfully, because the
behaviour of Future.sequence is such that it will fail as soon as
any of Futures fails, and ongoing futures will be interrupted. The
only way for sequence of futures to succeed is that all futures
succeed.