Persist kokelas to be done in batches
If all kokelakset are given to KokelasPersister,
they will be all pushed into mailbox of arvosana actor,
which will cause problems. One approach was to throttle
the requests and give the arvosana actor more time for processing.
This solution does not throttle but sends kokelakset in
small batches to KokelasPersiter, and will not give the next
batch before the current one is finished.